requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
black==24.4.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.40.0
cryptography==42.0.8
distlib==0.3.8
dj-database-url==2.2.0
Django==4.2.13
django-cors-headers==4.4.0
django-daraja==1.3.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
gunicorn==22.0.0
idna==3.7
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pbr==6.0.0
pillow==10.3.0
pipenv==2024.0.1
pipenv-to-requirements==0.9.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3